configured linting
