+++ .eslintignore
@@ -1,0 +1,3 @@
+/dist/
+/test/
+/coverage/
