+++ .prettierrc.js
@@ -1,0 +1,7 @@
+module.exports =  {
+  semi:  false,
+  trailingComma:  'all',
+  singleQuote:  true,
+  printWidth:  120,
+  tabWidth:  2,
+};
