--- api/v1/v1.ts
@@ -1,10 +1,0 @@
-import { Router } from "express";
-import { logger } from '../../utils/logger';
-
-export const routerV1 = Router();
-
-routerV1.use(logger)
-
-routerV1.get('/',(req, res)=> {
-  res.send('Welcome to Authors Haven')
-})
