+++ package.json
@@ -5,16 +5,35 @@
   "main": "server.ts",
   "scripts": {
     "start-server": "tsc && node build/server.js",
-    "start": "nodemon -e ts -w ./ -x npm run start-server"
+    "start": "nodemon -e ts -w ./ -x npm run start-server",
+    "lint": "tsc --noEmit && eslint 'src/*/**/*.{js,ts,tsx}' --quiet --fix"
+  },
+  "lint-staged": {
+    "*.{js,ts}": [
+      "git add"
+    ]
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@types/morgan": "^1.7.37",
     "express": "^4.17.1",
-    "morgan": "^1.9.1"
+    "morgan": "^1.9.1",
+    "typescript": "^3.6.2"
   },
   "devDependencies": {
-    "@types/express": "^4.17.1"
+    "@types/express": "^4.17.1",
+    "@typescript-eslint/eslint-plugin": "^2.2.0",
+    "@typescript-eslint/parser": "^2.2.0",
+    "eslint": "^6.3.0",
+    "eslint-config-prettier": "^6.2.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.17.3",
+    "eslint-plugin-node": "^9.1.0",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "prettier": "^1.18.2",
+    "lint-staged": "^8.1.5"
   }
 }
