+++ src/utils/logger.ts
@@ -1,0 +1,3 @@
+import morgan from 'morgan'
+
+export const logger = morgan('dev')
