--- utils/logger.ts
@@ -1,3 +1,0 @@
-import morgan from "morgan";
-
-export const logger = morgan('dev');
